Restructuring Wolf Sheep Model
Start restructuring WolfSheep into a subdirectory
