+++ examples/WolfSheep/model/model.py
@@ -11,14 +11,12 @@ Replication of the model found in NetLogo:
 
 import random
 from collections import defaultdict
-
-from mesa import Model, Agent
+from mesa import Model
 from mesa.space import MultiGrid
 from mesa.time import RandomActivation
 from mesa.datacollection import DataCollector
 
-from RandomWalk import RandomWalker
-
+from agents import Sheep, Wolf, GrassPatch
 
 class WolfSheepPredation(Model):
     '''
@@ -140,119 +138,6 @@ class WolfSheepPredation(Model):
                 self.schedule.get_breed_count(Wolf))
             print('Final number sheep: ',
                 self.schedule.get_breed_count(Sheep))
-
-class Sheep(RandomWalker, Agent):
-    '''
-    A sheep that walks around, reproduces (asexually) and gets eaten.
-
-    The init is the same as the RandomWalker.
-    '''
-
-    energy = None
-
-    def __init__(self, grid, pos, moore, energy=None):
-        super().__init__(grid, pos, moore)
-        self.energy = energy
-
-    def step(self, model):
-        '''
-        A model step. Move, then eat grass and reproduce.
-        '''
-        self.random_move()
-        living = True
-
-        if model.grass:
-            # Reduce energy
-            self.energy -= 1
-
-            # If there is grass available, eat it
-            this_cell = model.grid.get_cell_list_contents([self.pos])
-            grass_patch = [obj for obj in this_cell 
-                           if isinstance(obj, GrassPatch)][0]
-            if grass_patch.fully_grown:
-                self.energy += model.sheep_gain_from_food
-                grass_patch.fully_grown = False
-
-            # Death
-            if self.energy < 0:
-                model.grid._remove_agent(self.pos, self)
-                model.schedule.remove(self)
-                living = False
-
-        if living and random.random() < model.sheep_reproduce:
-            # Create a new sheep:
-            if model.grass:
-                self.energy /= 2
-            lamb = Sheep(self.grid, self.pos, self.moore, self.energy)
-            model.grid.place_agent(lamb, self.pos)
-            model.schedule.add(lamb)
-
-
-class Wolf(RandomWalker, Agent):
-    '''
-    A wolf that walks around, reproduces (asexually) and eats sheep.
-    '''
-
-    energy = None
-
-    def __init__(self, grid, pos, moore, energy):
-        super().__init__(grid, pos, moore)
-        self.energy = energy
-
-    def step(self, model):
-        self.random_move()
-        self.energy -= 1
-
-        # If there are sheep present, eat one
-        x, y = self.pos
-        this_cell = model.grid.get_cell_list_contents([self.pos])
-        sheep = [obj for obj in this_cell if isinstance(obj, Sheep)]
-        if len(sheep) > 0:
-            sheep_to_eat = random.choice(sheep)
-            self.energy += model.wolf_gain_from_food
-
-            # Kill the sheep
-            model.grid._remove_agent(self.pos, sheep_to_eat)
-            model.schedule.remove(sheep_to_eat)
-
-        # Death or reproduction
-        if self.energy < 0:
-            model.grid._remove_agent(self.pos, self)
-            model.schedule.remove(self)
-        else:
-            if random.random() < model.wolf_reproduce:
-                # Create a new wolf cub
-                self.energy /= 2
-                cub = Wolf(self.grid, self.pos, self.moore, self.energy)
-                model.grid.place_agent(cub, cub.pos)
-                model.schedule.add(cub)
-
-
-class GrassPatch(Agent):
-    '''
-    A patch of grass that grows at a fixed rate and it is eaten by sheep
-    '''
-
-    def __init__(self, fully_grown, countdown):
-        '''
-        Creates a new patch of grass
-
-        Args:
-            grown: (boolean) Whether the patch of grass is fully grown or not
-            countdown: Time for the patch of grass to be fully grown again
-        '''
-        self.fully_grown = fully_grown
-        self.countdown = countdown
-
-    def step(self, model):
-        if not self.fully_grown:
-            if self.countdown <= 0:
-                # Set as fully grown
-                self.fully_grown = True
-                self.countdown = model.grass_regrowth_time
-            else:
-                self.countdown -= 1
-
 
 class RandomActivationByBreed(RandomActivation):
     '''
